pyproject.toml
[project]
name = "mangarepacker"
version = "0.0.1"
description = "A Qt GUI application to manage manga (and comics) in .cbz format"
authors = [
    { name = "Alessandro Manente", email = "manente.alessandro@gmail.com"}, 
    {name = "Edoardo Napoli", email = "edoardo.napoli54@gmail.com"},
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = ["pyqt6 (>=6.9.0,<7.0.0)", "psutil (>=7.0.0,<8.0.0)", "pillow (>=11.2.1,<12.0.0)", "rarfile (>=4.2,<5.0)", "pyunpack (>=0.3,<0.4)", "patool (>=4.0.1,<5.0.0)", "numpy (>=2.2.6,<3.0.0)", "jsonpickle (>=4.1.0,<5.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
pre-commit = "^4.2.0"
pyinstaller = "^6.13.0"

